Reverse order of entries in CHANGELOG.md
The common practice is to order entries in the changelog in descending order so that the most recent changes appear at the top. When a new major release is coming out and you need to check it for breaking changes, it is much easier to open the changelog and see it immediately, rather than scrolling all the way down.